fix: header responsive design, add contact, routes and detail design fix
