+++ src/components/common/footer.tsx
@@ -49,46 +49,46 @@ export default function Footer() {
                 </div>
 
                 {/* Right: Subscribe */}
-                <FooterForm>
-                    <h4 className="text-sm font-semibold mb-2">Subscribe</h4>
-                    <Form {...form}>
-                        <form
-                            onSubmit={form.handleSubmit(onSubmit)}
-                            className="flex flex-between justify-center items-start gap-2 "
-                        >
-                            <FooterFormField
-                                control={form.control}
-                                name="email"
-                                type="email"
-                                aria-required="true"
-                                autoComplete="email"
-                            />
-                            <Button
-                                type="submit"
-                                aria-busy={isPending}
-                                disabled={isPending}
-                                variant="secondary"
-                                size="lg"
-                                className="rounded-md text-lg px-6 py-3 shadow transition transform duration-200 hover:bg-secondary/80"
-                            >
-                                {isPending ? (
-                                    <>
-                                        <Spinner size="small" />
-                                        <span aria-hidden="true">Subscribing...</span>
-                                    </>
-                                ) : (
-                                    "Subscribe"
-                                )}
-                            </Button>
-                        </form>
-                    </Form>
-                    <p className="text-xs text-gray-500 mt-2 text-right">
-                        By subscribing you agree to with our{" "}
-                        <Link href="/privacy-policy" className="underline hover:text-black">
-                            Privacy Policy
-                        </Link>
-                    </p>
-                </FooterForm>
+                {/* <FooterForm>
+          <h4 className="text-sm font-semibold mb-2">Subscribe</h4>
+          <Form {...form}>
+            <form
+              onSubmit={form.handleSubmit(onSubmit)}
+              className="flex flex-between justify-center items-start gap-2 "
+            >
+              <FooterFormField
+                control={form.control}
+                name="email"
+                type="email"
+                aria-required="true"
+                autoComplete="email"
+              />
+              <Button
+                type="submit"
+                aria-busy={isPending}
+                disabled={isPending}
+                variant="secondary"
+                size="lg"
+                className="rounded-md text-lg px-6 py-3 shadow transition transform duration-200 hover:bg-secondary/80"
+              >
+                {isPending ? (
+                  <>
+                    <Spinner size="small" />
+                    <span aria-hidden="true">Subscribing...</span>
+                  </>
+                ) : (
+                  "Subscribe"
+                )}
+              </Button>
+            </form>
+          </Form>
+          <p className="text-xs text-gray-500 mt-2 text-right">
+            By subscribing you agree to with our{" "}
+            <Link href="/privacy-policy" className="underline hover:text-black">
+              Privacy Policy
+            </Link>
+          </p>
+        </FooterForm> */}
             </div>
 
             {/* Divider */}
@@ -97,13 +97,13 @@ export default function Footer() {
             {/* Bottom Links + Copyright */}
             <div className="max-w-[95%] mx-auto flex flex-col lg:flex-row justify-between items-center text-sm text-gray-600 gap-2">
                 <div className="flex flex-wrap gap-4">
-                    <Link href="/privacy-policy" className="hover:underline">
+                    <Link href="#" className="hover:underline">
                         Privacy Policy
                     </Link>
-                    <Link href="/terms" className="hover:underline">
+                    <Link href="#" className="hover:underline">
                         Terms of Service
                     </Link>
-                    <Link href="/cookies" className="hover:underline">
+                    <Link href="#" className="hover:underline">
                         Cookies Settings
                     </Link>
                 </div>
