+++ src/constants/metadata.ts
@@ -1,40 +1,41 @@
 // src/constants/metadata.ts
 export const siteMetadata = {
-  baseUrl: "https://thinkstorm.app",
-  title: "ThinkStorm",
-  description:
-    "ThinkStorm is a unique collaborative platform to connect all tech professionals, learners and innovators.",
-  keywords: [
-    "ThinkStorm",
-    "Tech Professional",
-    "Software Engineer",
-    "Frontend Engineer",
-    "Backend Engineer",
-    "Full Stack Engineer",
-    "Software Developer",
-    "Frontend Developer",
-    "Backend Developer",
-    "Full Stack Developer",
-    "Developer",
-    "Designer",
-    "UI/UX Designer",
-    "Product Manager",
-    "Computer Science Student",
-    "Innovator",
-    "Entrepreneur",
-    "Collaborative Platform",
-    "IT Projects",
-    "IT Portfolio",
-    "Developer Portfolio",
-    "Student Projects",
-    "Learning Tech",
-    "Tech Stacks",
-  ],
-  author: "ThinkStorm",
-  ogImage: {
-    url: "/images/og-image.png",
-    width: 1200,
-    height: 630,
-    alt: "ThinkStorm",
-  },
+    baseUrl: "https://thinkstorm.app",
+    title: "ThinkStorm",
+    description:
+        "ThinkStorm is a unique collaborative platform to connect all tech professionals, learners and innovators.",
+    keywords: [
+        "ThinkStorm",
+        "Tech Professional",
+        "Software Engineer",
+        "Frontend Engineer",
+        "Backend Engineer",
+        "Full Stack Engineer",
+        "Software Developer",
+        "Frontend Developer",
+        "Backend Developer",
+        "Full Stack Developer",
+        "Developer",
+        "Designer",
+        "UI/UX Designer",
+        "Product Manager",
+        "Computer Science Student",
+        "Innovator",
+        "Entrepreneur",
+        "Collaborative Platform",
+        "IT Projects",
+        "IT Portfolio",
+        "Developer Portfolio",
+        "Student Projects",
+        "Learning Tech",
+        "Tech Stacks",
+    ],
+    author: "ThinkStorm",
+    ogImage: {
+        url: "/images/og-image.png",
+        width: 1200,
+        height: 630,
+        alt: "ThinkStorm",
+    },
+    email: "contact@thinkstorm.app",
 };
